requirements.txt
aiosqlite==0.20.0
alembic==1.14.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncio==3.4.3
asyncpg==0.30.0
backports.tarfile==1.2.0
bcrypt==4.2.1
billiard==4.2.1
black==25.1.0
build==1.2.2.post1
CacheControl==0.14.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crashtest==0.4.1
distlib==0.3.9
dnspython==2.7.0
dulwich==0.22.7
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.8
fastjsonschema==2.21.1
filelock==3.17.0
flake8==7.1.2
flower==2.0.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httptools==0.6.4
humanize==4.11.0
idna==3.10
importlib_metadata==8.6.1
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
keyring==25.6.0
kombu==5.4.2
Mako==1.3.8
MarkupSafe==3.0.2
mccabe==0.7.0
more-itertools==10.6.0
msgpack==1.1.0
mypy==1.15.0
mypy-extensions==1.0.0
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
pkginfo==1.12.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.50
pyasn1==0.6.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.12.1
redis==5.2.1
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
tomli==2.2.1
tomlkit==0.13.2
tornado==6.4.2
trove-classifiers==2025.1.15.22
types-pyasn1==0.6.0.20250208
types-python-jose==3.3.4.20240106
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
virtualenv==20.29.1
watchfiles==1.0.4
wcwidth==0.2.13
websockets==14.2
xattr==1.1.4
zipp==3.21.0
